Create ptraj specializations.  Split intersect into several files
