+++ Tests/Intersection_unit.cc
@@ -1,10 +1,9 @@
 //
 // Test intersections with KinKal objects
 //
-#include "KinKal/Trajectory/ParticleTrajectory.hh"
 #include "KinKal/General/ParticleState.hh"
 #include "KinKal/General/TimeRange.hh"
-#include "KinKal/Geometry/Intersection.hh"
+#include "KinKal/Geometry/Intersect.hh"
 #include "KinKal/Geometry/Cylinder.hh"
 #include "KinKal/Geometry/Disk.hh"
 #include "KinKal/Geometry/Annulus.hh"
