+++ Tests/ParticleTrajectoryTest.hh
@@ -8,7 +8,7 @@
 #include "KinKal/General/PhysicalConstants.h"
 #include "KinKal/Geometry/Cylinder.hh"
 #include "KinKal/Geometry/Disk.hh"
-#include "KinKal/Geometry/Intersection.hh"
+#include "KinKal/Geometry/ParticleTrajectoryIntersect.hh"
 
 #include <iostream>
 #include <cstdio>
@@ -264,7 +264,7 @@ int ParticleTrajectoryTest(int argc, char **argv) {
 
   Cylinder cyl(bnom,origin.Vect(),hlen,wlen);
   std::cout << "Test " << cyl << std::endl;
-  auto kc_inter = pieceIntersect(ptraj,cyl, ptraj.range(), 1.0e-8);
+  auto kc_inter = intersect(ptraj,cyl, ptraj.range(), 1.0e-8);
   std::cout << "KinematicLine Cylinder Intersection status " << kc_inter.flag_ << " position " << kc_inter.pos_ << " time " << kc_inter.time_ << std::endl;
 
   //
